Turn off Tomcat META-INF/MANIFEST.MF scanning
With a recent Tomcat, 9.0.13, loads of warnings about Tomcat being
unable to scan a JAR file which I never had reference nor did exist on
the file system:

```text
org.apache.tomcat.util.
scan.StandardJarScanner.scan Failed to scan
[file:/opt/tomcat-engine1/escenic/lib/serializer.jar] from classloader
hierarchy
```

It turned out, with Tomcat 8.0.41 or thereabouts, it has started to
scan all manifests for other JARs and add these to the classpath. If
any of these cannot be found, a big fat warning and stack trace is
added to the Tomcat standard out log.

For a typical Content Engine installation, this meant hundreds of
uninteresting log messages on each startup of the app server.