VOSA-617: Added "ece apply-conf" command
This new feature is here to simplify the creation (and possibly
maintenance) of a Content Store. Together with an .ini file which
describes the OUs, publication types, publications, tags, shared
resources, and so on, the new ece apply-conf will create them all. If
something is already there, it is "refreshed", i.e. changed files are
uploaded, unchanged files are left alone.

Hopefully, this can replace a lot of bespoke "create publication" and
similar things left and right.